requirements.txt
attrs==18.2.0
blinker==1.4
click==8.0.1
Flask==2.0.1
Flask-Mail==0.9.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pdfkit==0.6.1
pdflatex==0.1.3
python-dotenv==0.19.0
Werkzeug==2.0.1
